Fixed a bug in equalsFloat when maxDiff = 0.
It was always returning false due to 'rounded' being NaN.
The consequence was that 'hasChanged' in huelight was always returning
false when the light was in xy color mode, so manual color changes
were not detected. Color temperature changes were correctly detected,
which is probably the reason why this bug was not found earlier.